+++ packages/loggers/logger-with-callback/lib/index.ts
@@ -1,0 +1,1 @@
+export { LoggerWithCallbackFactory } from './loggerWithCallbackFactory';