+++ packages/tokens/vitest.config.ts
@@ -1,0 +1,18 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    globals: true,
+    coverage: {
+      all: true,
+      reporter: ['cobertura', 'lcov', 'text', 'text-summary'],
+      provider: 'istanbul',
+      include: ['src'],
+    },
+    poolOptions: {
+      forks: {
+        singleFork: true,
+      },
+    },
+  },
+});