+++ packages/bitcoin-utxo-selection/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @rosen-bridge/bitcoin-utxo-selection
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/abstract-logger@2.0.1
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/abstract-logger@2.0.0
+
 ## 0.2.0
 
 ### Minor Changes