+++ packages/boxes/rwt-repo/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @rosen-bridge/rwt-repo
 
+## 2.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/abstract-logger@2.0.1
+
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/abstract-logger@2.0.0
+
 ## 2.0.2
 
 ### Patch Changes