+++ packages/ergo-box-selection/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @rosen-bridge/ergo-box-selection
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/abstract-logger@2.0.1
+
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/abstract-logger@2.0.0
+
 ## 0.5.0
 
 ### Minor Changes