+++ packages/rosen-extractor/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @rosen-bridge/rosen-extractor
 
+## 6.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/abstract-logger@2.0.1
+
+## 6.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/abstract-logger@2.0.0
+
 ## 6.2.0
 
 ### Minor Changes